Update template.md
A few of the metadata elements are missing descriptive info in the Metadata section. I can't get internal links to work in the "internal links" section. I didn't do much to the Code section. After this has been committed to Master, I'll send a note to Robin to review and update the Code section.