Updated all over again.
Please - I am conscious of choosing the 'past-tense' words like 'were' instead of 'are' as it's describing the change in usability of an already released feature. It took me lot of time to pull it all together again. I sincerely request - to kindly connect with Engineering team when CP team has to modify majority of the content structure. Also, I am not sure how to fix the metadata - as that seems to have happened during migration. I would request CP team to kindly help me with this one. Lastly, screenshots are there to explain the context and not all users would always prefer it to be only bullet points about how-to as such.