Incorporated Tonya's feedback.
In the last section, add the UI navigation (in words) on how to do this. Add something like “To do XX, go to XX”. Right now there is one sentence and screenshot, neither of which show the user how to select properties to update products from the Retail product category form, the text just says "you can use this..." feature. 
I don't recommend using "Enhanced..." in the title, because this feature is going to become commonplace and no longer be "enhanced." Instead, consider changing the title to something that will continue to remain fresh, such as "Manage Retail product categories and products."