Update configure-mobile-devices-warehouse.md
"The option can be chosen under the **Directed by** field in the **Mobile device menu items** page" added.
It is needed to add that sentence so customers can look for "Directed by" in that page and go to that section. Helptext on "Directed by" field will be included with a link to this docs page, so we want customers to find it easily.