Update app-sealing-faq.md
Added some PU distinguishing...
Developers should not depend on these classes and methods remaining unchanged.
Calls to classes and methods with the **InternalUseOnlyAttribute** will result in compiler warnings. In platform update 20 to platform update 24, targeting classes and methods with **InternalUseOnlyAttribute** using Chain of Command will result in compiler errors, but in platform update 25 and onwards we plan to make these compiler warnings as well.