Update synchronize-project-list.md
Please note that "Fin and Ops" should not be changed to Finance and Operations, as it is referring to the released template names - e.g. "Work Orders with Project (Field Service to Fin and Ops)" we are allowed to use Fin and Ops here. Main reason being that the template name has a max number characters that would be excited with the full name.