Updating Process offline loyalty transactions instructions
Process name is: Post earned loyalty points in batches.  Description is: Run this process to update loyalty cards so that they include transactions that were processed offline. This process applies only if the **Post earned points in batches** check box is selected on the **Retail shared parameters** page, so rewards can be earned offline.  Page name is: Post earned loyalty points in batches