Update deprecated-features.md
Removing "(If and only if you need to repair data corruption caused by a previous rename of the primary key of a released product, please contact Microsoft Support to request temporary removal of this restriction.)"

Customers still contact us to see if they can rename for other purposes because they see this. Removing this option - if they had corrupted data because of this, then they have had a lot of time (since 10.0.0 until now 10.0.9) to do this. So no longer stating this publically anymore to save the trouble of people asking this for other purposes