note to changes made in the web editor
We had a customer that had extensively customized an embedded report via the web editor and they went back to then edit over those changes on a Dev server believing their changes from the web would be available to the PBIX to then edit in PowerBI Desktop on a dev environment.  This is not the case and they were quite upset that the feature to SAVE customizations didn't allow them to access those or retrieve the modified PBIX file.  The likely missed the other section where we do call out the fact that customizations are not exportable so it may be ideal if we call it out again in a Note as edited.  Had they known that they wouldn't have invested as much time in editing in the web form and just done it from the dev machine with PowerBI Desktop so they could manage those edits going forward with change control.