Update troubleshoot-masterplanning.md
Updated based on request from Smitha
Area: Master planning > Master plans
I removed the following as this is a question and also master planning don't respect this: 
##  Is it possible to set a maximum quantity for a warehouse?
The maximum quantity for a warehouse can be set by setting the location stocking limits for the warehouse. To do that goto: Warehouse management > Setup >  Warehouse > Location stocking limits.

The purpose of Master planning is to ensure that you have supply to meet demand. There is no functionality to change destination after a certain maximum value nor functionality to automatically transfer material when it is above the maximum set via the Location stocking limits.
		
**Resolution/Fix**
That's because the delivery address, site and warehouse doesn't automatically get changed at the line level either - you'll need to update that yourself