Update release-product-structure.md
Removing the below as it is not true. It is optional both for engineering and standard products. 

For engineering products, the release policy is assigned in the engineering product category, and the field is mandatory. For standard products, the policy is assigned to the shared product, and the field is optional.