Removal of users from role as specified more clearly
Because I am doing an accreditation for Dynamics for a public customer, I have a requirement, that our documentation clearly show both how to add and remove user security. So, I have updated this page to be able to be specific enough to get the accreditation approved at the customer.