Functionality works per-line level
It is more correct to describe "Requirement for inventory reservation" as working per order **line** level, not for the whole order. Because by itself this feature does not try to ensure that all order lines have the reservation or cross docking. This feature checks just the quantities on the given lines being released. It means, one could open "Release to warehouse" form and manually select to release just a single order line from there. In that case, this feature would make no attempt at all to see what other lines order has and if the reservations are done there.

This is unlike the "Fulfillment policies" functionality which indeed works on the order level.