Update emea-sau-e-invoices.md
Customer testing their E-Invoicing in Saudi Arabia was facing some issues and our analysis shows that they had done an incorrect setup because of which the invoice generated by them had incorrect fields. This caused the validation errors while doing compliance verification of their E-Invoicing solution.
The issue was “The sales tax code was not same as external code”. This condition we have not mentioned in the Microsoft documentation, hence this change in docs aims to help users and avoid the errors.