Delete Important notice
This erroneous behavior was fixed in scope of this issue: Details for issue 770099 (dynamics.com), so that the LCS EU end point is corrected as well as LCS auth logic was moved to LCS parameters tab activation where it is actually needed.
With this fix, the system behavior is getting clearer for e-Invoicing scenarios and LCS related errors would not pop up since they are completely irrelevant here. We may want to even delete this Important notice from e-Invoicing documentation.