Add missing "t" in data-management-api.md
**Incorrect Paragraph: **
There is an async version of this API ImportFromPackagePreviewAsync. The specifications are the same. I will be required to capture the execution id returned and the later call the GetExecutionSummaryStatus API to determine when the execution has completed.


**Correct Paragraph: **
There is an async version of this API ImportFromPackagePreviewAsync. The specifications are the same. It will be required to capture the execution id returned and the later call the GetExecutionSummaryStatus API to determine when the execution has completed.