Update security-architecture.md
System Admin-->Inquiries-> User Log
Reason for deprecation/removal:	The User Log is a legacy page that was built for the older client/server architecture. The information on this page is not always accurate and can be misleading.
Replaced by another feature:	In Microsoft Dynamics 365 finance and operations apps, this information is captured in telemetry and LCS has details.


Removed the below as the User log form reference:
Auditing
For more information, see [https://learn.microsoft.com/en-us/dynamics365/fin-ops-core/dev-itpro/lifecycle-services/user-logins)
Auditing of user sign in and sign out is enabled, which means that the system logs when a user signs in or out of the application. A sign out is logged even if the user's session expires or ends.

A system administrator or security administrator can access the audit logs by going to  the **User log** page (**System administration** > **Inquiries** > **User log**).

Replaced with Auditing user logins:
For more information, see [https://learn.microsoft.com/en-us/dynamics365/fin-ops-core/dev-itpro/lifecycle-services/user-logins)