Update business-continuity-disaster-recovery.md
removed this statement : If Financial Reporting is critical for a customer during the disaster, the customer can request restoration of the service to Microsoft through a support ticket.
The document mentions that FR isn’t available while the env. is failed over to the secondary region (a temporary state). However, since the switch from IaaS to self-service the highlighted statement is no longer true. There’s no currently a procedure that allows restoring FR during the failover. In that case, they need to wait until the primary region is restored.