Update emea-fra-VAT-declaration-preview-France.md
The Note about the feature, Use application specific parameters from previous versions of ER formats, is not relevant any longer as the feature does not exists any longer in version 10.0.40 (10.0.1935.69) with the name: "Use application specific parameters from previous versions of ER formats"

NB that the Note is exists in a lot of articles in the docs. I have just suggested the change for France but I think the same change should be done everywhere the note exists.