Update clean-up-data.md
Automatic cleanup of data more than 90 days retention. This feature work is in progress with some changes. The earlier implementation is on hold due to customer feedback and it is not the correct update for now in public doc. We will update this data once we have the feature with changes rolled out. There is no requirement of customer action here.