Update thailand.md - Removed reference to Microsoft support.
The field System parameters > Sequence number is required for any new legal entities in Thailand localisation. Yet, this field plays no purposes other than showing in "About" form.

Yet, our article on MS Learn is referring to Microsoft Support for the valid value of this field. This is causing an over-emphasised importance of this field.

Microsoft Support does not know this number and as I checked in our internal resources, there's no one really clear of which number that is deemed being valid.